verifier: split into utils/ + rules/, drop sweep + check_renderers
Folder layout
- utils/ — engine modules (probe, inspect_rules, inspect, hf_parity)
- rules/ — data files (inspect_rules.yaml, default_prompts.json)
- top level — CLI entry points only (cli, serve, triage, spinup_deployment)
- SKILL.md drives the test flow; the package exposes utilities, not
  many overlapping pre-packaged scripts.

Removals
- sweep.py: parallel verdict heuristic, not the inspect_rules.yaml
  source of truth — drop.
- check_renderers.py: redundant with triage's pre-flight ping; SKILL.md
  describes how to verify reachability with a one-prompt triage run.

User-facing CLI paths unchanged
- python -m training.verifier render | inspect
- python -m training.verifier.serve
- python -m training.verifier.triage
- python -m training.verifier.spinup_deployment

Tests updated; 9 passed + 1 xfail (existing MiniMax xfail).
