+++ training/verifier/rules/inspect_rules.yaml
